[android] merge enumerate_backends.cpp to entry_android.cpp
