+++ android/app/src/main/cpp/entry_android.cpp
@@ -1,6 +1,7 @@
-#include <opencl/cl_peak.h>
 #include <common/peak.h>
 #include <common/options.h>
+#include <common/inventory.h>
+#include <version.h>
 #include "logger_android.h"
 #include "jni_entry.h"
 
@@ -9,6 +10,9 @@
 #endif
 #ifdef ENABLE_CPU
 #include <cpu/cpu_peak.h>
+#endif
+#ifdef ENABLE_CPU
+#include <opencl/cl_peak.h>
 #endif
 
 #define RECORD_METRIC_CALLBACK "record_metric_callback_from_c"
@@ -27,6 +31,26 @@ static void wireLoggerToJni(LoggerAndroid *lg, JNIEnv *jniEnv, jobject jObj, jcl
                                                  DEVICE_INFO_CALLBACK,
                                                  "(Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;"
                                                  "Ljava/lang/String;Ljava/lang/String;Ljava/lang/String;)V");
+}
+
+static std::vector<BackendInventory> enumerateAllBackends(const CliOptions &opts)
+{
+    std::vector<BackendInventory> out;
+
+#ifdef ENABLE_VULKAN
+    if (!opts.skipVulkan)
+        out.push_back(vkPeak::enumerate());
+#endif
+#ifdef ENABLE_OPENCL
+    if (!opts.skipOpenCL)
+        out.push_back(clPeak::enumerate());
+#endif
+#ifdef ENABLE_CPU
+    if (!opts.skipCpu)
+        out.push_back(CpuPeak::enumerate());
+#endif
+
+    return out;
 }
 
 jint JNICALL Java_kr_clpeak_BenchmarkRepository_launchClpeak(JNIEnv *_jniEnv,
@@ -83,7 +107,12 @@ jint JNICALL Java_kr_clpeak_BenchmarkRepository_launchClpeak(JNIEnv *_jniEnv,
     return clStatus;
 }
 
-#include <version.h>
+extern "C" JNIEXPORT jstring JNICALL
+Java_kr_clpeak_BenchmarkRepository_nativeEnumerateBackends(JNIEnv *env, jobject)
+{
+    CliOptions opts;
+    return env->NewStringUTF(inventoryToJson(enumerateAllBackends(opts)).c_str());
+}
 
 jstring Java_kr_clpeak_AboutBottomSheet_nativeGetVersion(JNIEnv *jniEnv, jobject _jObj)
 {
